Rename foreman package.
At this point, saying "local" is redundant.  And it's starting to look wholy possible that most of the planning parts being written here will actually be general to both farm and non-farm situations after all.
